App Version 2.0 (Working order)
I have added functionality and logic to several buttons within the Kiosk class, and those are the ScanItemsBtn has functionality inside that communicates with a Jtextfield, JList and JLabel, where by the customer will enter the name of the item they want to buy (which is inside the stock.txt) and it will display the totalprice of those items in a JLabel.

The CashPaymentBtn has functionality whereby after the customer has the totalprice of the items, they will enter into another textfield the amount of money they have to buy the items, and press cash JButton and it will display their change above.

The CardPaymentBtn now link to another form where the user will login to their bank and see their balance, however the logic behind this has not been completed yet.

I have made a new class called Dataloader which holds the logic for reading both admin and stock text files, allowing the scanner to read the files using the separator to locate each piece of data between the break point. There are also two constructors, one for my stockdata and AdminData which contain more logic allowing its functionality inside the form, thus allowing the Jbuttons to work as they do.

Finally we have the stockdata and admindata classes which hold our variables being used inside our constructors, and also our getters and setters.