App Version 2.0 (Modified)
I have began to add functionality within the Dataloader class by adding a constructor for the bankdata by adding an ArrayList which holds logic for a method called loadbankdata, inside that constructor is an try catch again this time adding a new user to the text file bankinformation.txt, this works by having another constructor called save which applies this logic.

The array and constructor has been added to the BankVerificationScreen class and I will continue to apply more functionality to this form allowing the user to check in successfully.

Also a bankdata class has been added where varaibles for name, age and balance and if the account is platinum, along with getters and setters.