App Version 3.2 Re-worked
AdminLogin wasn't functioning properly due to incorrect formatting, removed static array and replaced with exsisiting arraylist which was located in Dataloader but wasn't being used...

With the implementation of the ArrayList and setter for said ArrayList, I was able to reformat by AuthenticateUser method where the logic is placed. ]

Also, made app look colourful.